requirements.txt
fastapi
uvicorn[standard]
qdrant-client
fastembed
Pillow
boto3
python-dotenv
mysql-connector-python
deep-translator
bcrypt
sentence-transformers
